Insert new games into db and migration stuff

+++ database/migrations/2020_03_11_212131_create_games_table.php
@@ -17,7 +17,7 @@ class CreateGamesTable extends Migration
             $table->bigIncrements('id');
             $table->string('name');
             $table->string('type');
-            $table->integer('price');
+            $table->double('price', 8, 2);
             $table->timestamps();
         });
     }
